Small fixes & typos (#7)
* Fix: Restarting the game keeps perks from the previous run

The finalPerks array wasn't getting reset. Now it is.

* Bathroom grammar update

Added a "and" to the final entry in the digested prey list

* Typo fixes

Corrected some typos

* Made digestMessages a class member instead of a function value

Since the messages don't change, I moved them to a more permanent
location.  This will improve performance by an unnoticeable amount, but
I think it makes the code a bit cleaner.

* Added function to truncate excessive decimal places in output

Truncates numbers to 2 decimal places

* Typo fix & variable rename

Changed "streach" to "stretch"

* Digestion/time refactor

Ok! This one's a pretty big change, I won't be offended if you don't
want to pull this one. :-)

The basic idea was to centralize time-passage functionality so that
anybody can just call "passTime" and not have to worry about
implementing duplicate digestion/stretching/recovery logic.

Along with that, the digestion, stretching, and health changes were
split into separate functions. This is mostly for cleanliness and
readability, to make it easier to make changes or find bugs later.

I tried to keep most of the functionality the same. Where things were
changed, I put a "$design" comment.

- Moved digestion, stretching, health, and time tracking into separate
functions in MyPlayerObject
- Removed digestion handling in doFoodComa(), it uses passTime() now.
- Moved "enter food coma" check from passTime to movePlayer.